+++ src/routes/index.ts
@@ -11,6 +11,7 @@ import breadcrumbs from './breadcrumbs/breadcrumbs';
 import catalog from './catalog/catalog';
 import content from './content/content';
 import content_rules from './content_rules/content_rules';
+import contextnavigation from './contextnavigation/contextnavigation';
 import controlpanels from './controlpanels/controlpanels';
 import database from './database/database';
 import email from './email/email';
@@ -45,6 +46,7 @@ export default [
   ...breadcrumbs,
   ...catalog,
   ...content_rules,
+  ...contextnavigation,
   ...controlpanels,
   ...database,
   ...email,